# docs/swagger.yaml
basePath: /
definitions:
...
  dto.GetHistory:
    properties:
      booking_code:
        type: string
      cinema_logo:
        type: string
      cinema_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      movie_id:
        type: integer
      payment_status:
        type: string
      poster_url:
        type: string
      show_date:
        type: string
      show_time:
        type: string
      ticket_count:
        type: integer
      title:
        type: string
      total_price:
        type: integer
    type: object
  dto.GetProfile:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      loyalty_points:
        type: integer
      phone_number:
        type: string
      profile_image:
        type: string
      role:
        type: string
    type: object
...
  dto.PaginationMeta:
    properties:
      next_page:
        type: string
      page:
        type: integer
      prev_page:
        type: string
      total_page:
        type: integer
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      msg:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.UpdateProfileResponse:
    properties:
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      profile_image:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: Back End using go with gin.
  title: Tickitz Back End
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get current user profile information (Requires user token)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetProfile'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - user
  /user/history:
    get:
      consumes:
      - application/json
      description: Get user's order history (Requires user token)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GetHistory'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get order history
      tags:
      - user
  /user/password:
    patch:
      consumes:
      - application/json
      description: Change user password (Requires user token)
      parameters:
      - description: Password Update Body
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update user password
      tags:
      - user
  /user/profile:
    patch:
      consumes:
      - multipart/form-data
      description: Update user profile information including image upload (Requires
        user token)
      parameters:
      - description: First Name
        in: formData
        name: first_name
        type: string
      - description: Last Name
        in: formData
        name: last_name
        type: string
      - description: Phone Number
        in: formData
        name: phone_number
        type: string
      - description: Profile Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"